+++ JiraPS/Public/Invoke-JiraIssueTransition.ps1
@@ -16,16 +16,9 @@
         $Fields,
 
         [Parameter( ParameterSetName = 'AssignToUser' )]
-        [ValidateNotNullOrEmpty()]
-        [ValidateScript(
-            {
-                if ($_ -is [string] -and [string]::IsNullOrWhiteSpace($_)) {
-                    throw "The -Assignee value cannot be a whitespace-only string. Use -Unassign to remove the assignee."
-                }
-                $true
-            }
-        )]
-        [Object]
+        [ValidateNotNull()]
+        [AtlassianPS.JiraPS.UserTransformation()]
+        [AtlassianPS.JiraPS.User]
         $Assignee,
 
         [Parameter( ParameterSetName = 'Unassign' )]